catalog-backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlmodel==0.0.22
asyncpg==0.31.0
sqlalchemy[asyncio]==2.0.36

# Migrations
alembic==1.14.0

# Configuration
python-dotenv==1.0.1
pydantic-settings==2.7.1

# Utilities
python-slugify==8.0.4

